Add defaults to collect logs for image building.
Due to how we're using this role in our CI (calling it from a publisher
as a post build action) we are unable to pass ansible variables to the role.

1. The role is included in the requirements.txt so that collect-logs[-internal].sh
   is unpacked to $WORKSPACE/tripleo-quickstart/ci-scripts as part of quickstart.sh
   invocation #1 (the actual oooq run)

2. As a post build action, the script is run, which wraps a second invocation to
   quickstart.sh's using the retain inventory flag so the static hosts file isn't nuked
   and can be used by the collect logs role.  This allows it to be used as a post build
   action, which is run irrespective of pass/fail of the job.

Since THAT script does not (yet/now) take args (such as a number of additional log files
that we know will exist (e.g. image build logs), we need to add them to defaults.  When they
are not present the role ignores their absense.

In the future it would be great to have something a bit simpler and allowing for customization
via ansible variables without either property files (even more complicated) or some other
mechanism.  For now this is the KISS solution.

Change-Id: Ib67351d7856bd0ff8638a8d3af07fdaa390f721f