Explicitely pass RSYNC_PASSWORD when uploading logs to ci.c.o artifacts
The rdo-jenkins.ci.centos.org jenkins slave has an implicit export
of RSYNC_PASSWORD located in .bash_profile which is sourced when
spawning a new shell.
However, this variable is already available and configured on the
jenkins slave (in the ci.c.o web interface).

When setting up and testing new slaves without that .bash_profile
export, the RSYNC_PASSWORD variable was not available in that context.

We noticed the issue when weirdo jobs would be able to upload
artifacts while the quickstart jobs failed.

Change-Id: I6aababdf48fe8a5b19acaecb1b8427a68591973f