Update patch set 1
Patch Set 1: Code-Review+1

yep this is correct.

The ansible shell module creates a new shell, so we dont have the env vars inherited from jenkins unless we put them in the new shell explicitly.

The reason this works on the existing slave is that we "explicitly" inject it into **all** new shells via .bash_profile. It is better to make it like this so we don't have to copy that .bash_profile around.

Patch-set: 1
Label: Code-Review=+1