Split role call into logical segements
In preperation for integrating tripleo-documentor funtionality into
tripleo-collect-logs, it makes sense to break apart the collect and
publish sections of the role into includes. This keeps things neat and
distinct.

Change-Id: I165dc823ae0dd489a085c1bb7c00871db2d988ba

Split role call into logical segements

In preperation for integrating tripleo-documentor funtionality into
tripleo-collect-logs, it makes sense to break apart the collect and
publish sections of the role into includes. This keeps things neat and
distinct.

Change-Id: I165dc823ae0dd489a085c1bb7c00871db2d988ba