Update patch set 2
Patch Set 2: Code-Review+2

This is the point where I get a bit uneasy about the amount of data collected by the role. We can mostly assume that the passwords in the config files are autogenerated but what if somebody uses their own certs and we gather it all and accidentally expose it?

We might want to add a disclaimer to the README that this can happen. With all this being said, we can merge this if it passes the gate.

Patch-set: 2
Label: Code-Review=+2